Split `Group` into modules
